spdmlib: fix cargo clippy warnings
- Allow clippy::duplicate_mod for intentional shared test module pattern
- Use struct initializers instead of field reassignment after Default::default()
- Derive Default for SpdmChunkStatus with #[default] attribute
- Replace if is_some()/unwrap() with if let Some()
- Remove unnecessary &mut references and casts
- Use .is_multiple_of() instead of manual modulo checks
- Replace assert_eq!(x, true/false) with assert!(x)/assert!(!x)
- Remove assert!(true/false), replace with panic!() or remove
- Use iterators instead of needless range loops
- Eliminate single-element loops
- Use unwrap_or_else instead of expect with format!
- Prefix unused variable with underscore
- Simplify boolean expressions

+++ spdmlib/src/message/chunk_get_test.rs
@@ -55,7 +55,7 @@ fn test_chunk_get_and_response_struct() {
     assert_eq!(u8_slice[4..6], [0, 0]);
     LittleEndian::write_u32(field_slice, chunk_size as u32);
     assert_eq!(u8_slice[6..10], field_slice[..]);
-    LittleEndian::write_u32(field_slice, large_message_size as u32);
+    LittleEndian::write_u32(field_slice, large_message_size);
     assert_eq!(u8_slice[10..14], field_slice[..]);
 
     context.chunk_context.transferred_size = 0;
